moved check_certificate_status into is own library.
