+++ lib/shelf_letsencrypt.dart
@@ -3,9 +3,10 @@ library shelf_letsencrypt;
 
 export 'src/certificates_handler_io.dart';
 export 'src/certs_handler.dart';
+export 'src/check_certificate_status.dart';
 export 'src/domain.dart';
 export 'src/domain_certificate_file_path.dart';
 export 'src/letsencrypt.dart';
+export 'src/logging.dart' show Logging;
 export 'src/pem_key_pair.dart';
 export 'src/security_context_builder.dart';
-export 'src/logging.dart' show Logging;
